archive/requirements.txt
dash-bootstrap-components==1.1.0
dash-daq==0.5.0
dash-extensions==0.1.3
dash-uploader==0.7.0a1
jupyter==1.0.0
notebook-as-pdf==0.5.0
numpy-stl==2.16.3
opencv-contrib-python==4.5.5.64
pathos==0.2.9
pip-chill==1.0.1
plumbum==1.7.2
ply==3.11
pymongo==4.1.1
pytest==7.1.2
python-dotenv==0.20.0
retrying==1.3.3
schedule==1.1.0
scikit-learn==1.1.1
seaborn==0.11.2
statsmodels==0.13.2
tensorflow==2.8.0